Migrate json schema
