+++ packages/json-schema/generated-defs/TypeSpec.JsonSchema.Private.ts
@@ -5,3 +5,6 @@ export type ValidatesRawJsonDecorator = (
   target: Model,
   value: Type
 ) => void;
+export type TypeSpecJsonSchemaPrivateDecorators = {
+  validatesRawJson: ValidatesRawJsonDecorator;
+};
